+++ src/lib/core/server.hpp
@@ -1,23 +1,9 @@
 #ifndef SRC_LIB_SERVER_HPP
 #define SRC_LIB_SERVER_HPP
 
-#include "repository.hpp"
-#include "event/event_notifier.hpp"
 #include "event/event_handler.hpp"
 #include "transport/listener.hpp"
-
-class ServerState {
-public:
-    EventNotifier &getEventNotifier();
-    ConnectionRepository &getConnectionRepository();
-    EventHandlerRepository &getEventHandlerRepository();
-
-private:
-    // EventNotifier はあんまり state っぽくない
-    EventNotifier notifier_;
-    ConnectionRepository connRepo_;
-    EventHandlerRepository handlerRepo_;
-};
+#include "server_state.hpp"
 
 class Server {
 public:
